code cleanup

+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-microsoft-module/src/main/java/org/apache/tika/parser/microsoft/onenote/fsshttpb/streamobj/StreamObjectHeaderStart32bit.java
@@ -42,7 +42,7 @@ public class StreamObjectHeaderStart32bit extends StreamObjectHeaderStart {
 
         if (length >= 32767) {
             this.length = 32767;
-            this.largeLength = new Compact64bitInt((long) length);
+            this.largeLength = new Compact64bitInt(length);
         } else {
             this.length = length;
             this.largeLength = null;
